backend/requirements.txt
flask
python-dotenv
supabase
flask-cors
pytest
pytest-mock
